+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 - Add workerpool address in configuration instead of reading from Scheduler `/workers/config` endpoint. (#607)
 - Set `0x0` as default value for Workerpool address and prevents startup if incorrectly configured. (#608)
 
+### Bug fixes
+
+- Use Result Proxy URL defined in deal parameters if any, fall back to scheduler default one otherwisea. (#613)
+
 ### Dependency Upgrades
 
 - Upgrade to `eclipse-temurin:11.0.24_8-jre-focal`. (#611)